First Pass At A* Search
Changed the way moving works. Now instead of inputting a piece to move, you simply input a direction you want the empty piece to move.

Also took an initial pass at the A* search algorithm. It appears something is going wrong with deciding on the best move.

TODO: Separate the best move decision and the actual movement and updating of the checked boards.